+++ ChangeLog.md
@@ -1,8 +1,8 @@
 # Changelog
 
 ## Version [4.0.0](https://github.com/cedx/cli/compare/v3.1.0...v4.0.0)
-- Added the `-Filter` parameter to the `ConvertTo-Encoding` cmdlet.
-- Added the `-Config` parameter to the `Install-Node` cmdlet.
+- Added the `-Filter` parameter to `ConvertTo-Encoding` and `Restore-MySqlTable` cmdlets.
+- Added the `-Config` parameter to `Install-Node` cmdlet.
 - Dropped the support of [PowerShell](https://learn.microsoft.com/en-us/powershell) applications in the `New-NssmService` cmdlet.
 - Fixed the base URL of [PHP](https://www.php.net) releases in the `Install-Php` cmdlet.
 - Ported the cmdlets of `Compression`, `MySql` and `Validation` modules to [C#](https://learn.microsoft.com/en-us/dotnet/csharp).